binary outcomes constrained step search
using Ipopt to find a local optimum of the constrained search problem
in the binary outcomes case. Removed some superfluous output. Need to
change the CV code to do stratified sampling in this case to make sure
at least one of each outcome case present in sample or else this will
crash.